Rename dependencies/events.js to dependencies/statistics.js
