+++ src/main/dependencies/index.js
@@ -20,7 +20,7 @@ import DIContainer from '../../app/di/jpex-container'
 import mysterionBootstrap from './modules/application'
 import bugReportingConfigBootstrap from '../../dependencies/bug-reporting'
 import bugReportingBootstrap from './modules/bug-reporting'
-import eventsBootstrap from '../../dependencies/events'
+import eventsBootstrap from '../../dependencies/statistics'
 import mysteriumClientBootstrap from './modules/mysterium-client'
 import mysteriumTequilapiBootstrap from '../../dependencies/mysterium-tequilapi'
 import proposalFetcherBootstrap from './modules/proposal-fetcher'
