+++ src/renderer/dependencies/index.js
@@ -20,7 +20,7 @@ import Vue from 'vue'
 import DIContainer from '../../app/di/vue-container'
 import bugReportingConfigBootstrap from '../../dependencies/bug-reporting'
 import bugReportingBootstrap from './modules/bug-reporting'
-import eventsBootstrap from '../../dependencies/events'
+import eventsBootstrap from '../../dependencies/statistics'
 import vueBootstrap from './modules/vue'
 import applicationBootstrap from './modules/application'
 import mysteriumTequilapiBootstrap from '../../dependencies/mysterium-tequilapi'
